Move ProtectedRoute component to a new directory; update import paths accordingly
